combine brib and barycentricfct to unitarybestapp
