+++ src/rexpi/__init__.py
@@ -1,5 +1,4 @@
-from .barycentricfcts import *
-from .brib import *
+from .unitarybestapprox import *
 from .cheb import *
 from .expimv import *
 from .pade import *
