+++ src/rexpi/expimv.py
@@ -1,66 +1,58 @@
 import numpy as np
-from scipy.special import jv
-
-#######################################################################
-#######################################################################
-#######################################################################
-
 
 ##########################################################################
 ##########################################################################
-##########################################################################
 
-def chebyshev(op,t,v,m):
+def evalr_product_scalar(z, poles, r0=1):
+    zs = np.asanyarray(z).ravel()
+    mvconj = lambda b : np.conj(zs)*b
+    mvSI = lambda s,b : b/(zs-s)
+    k = len(zs)
+    b = np.ones(k)
+    y = evalr_product(mvconj, mvSI, b, poles, r0=r0)
+    
+    if np.isscalar(z): return y[0]
+    y.shape = np.shape(z)
+    return y
+    
+def evalr_product(opconj, opSI, v, poles, r0=1):
     """
-    Clenshaw Algorithm
-    polynomial chebychev approximation return p(A)*v ~ exp(1j*t*A)*v  
-    op(v) = A*v applies operator on v,
-    the eigenvalues of A are assumed to be located in [-1,1]
-    m .. degree of p
-    """
-
-    cm1 = (1j)**m*jv(m,t)
-    dkp1 = cm1 * v
-    dkp2 = 0
-    for k in range(m-1,-1,-1):
-        ck = (1j)**k*jv(k,t)
-        dk = ck * v + 2*op(dkp1) - dkp2
-        if (k>0):
-            dkp2 = dkp1
-            dkp1 = dk
-    return dk - dkp2
-
-##########################################################################
-##########################################################################
-##########################################################################
-
-def evalr_product(op, opinv, v, poles, c0=None):
-    """
-    c0 = 1 for unitary best approximation with approximation error<2
+    opconj: v -> conj(A)*v
+    opSI: s, v -> inv(A-s)
+    A is skew-Hermitian (spectrum resides on imaginary axis)
     best approximation for skew-Hermitian matrix A with spectrum in i*[-w,w] 
     compute (prod_j (A-conj(sj)*I)*inv(A-sj*I))*v
     for poles sj in C without imaginary axis
+    assume r is symmetric s.t. constant factor is r0 = r(0)
+    r0 = 1 for unitary best approximation with approximation error<2
     """
-    if c0 is None:
-        n = len(poles)
-        c0 = (-1)**n
-    v2 = v.copy()
+    v2 = r0*v
     for s in poles:
-        v1 = opinv(s,v2)
-        v2 = op(v1)+np.conj(s)*v1
-    return c0*v2
+        v1 = opSI(s,v2)
+        v2 = opconj(v1)-np.conj(s)*v1
+    return v2
 
    
 ##########################################################################
 ##########################################################################
 
-def evalr_partialfraction(opinv, v, coef, poles, c0=None, v0=None):
+def evalr_partialfraction_scalar(z, coef, poles, rinf=None):
+    zs = np.asanyarray(z).ravel()
+    mvSI = lambda s,b : b/(zs-s)
+    k = len(zs)
+    b = np.ones(k)
+    y = evalr_partialfraction(mvSI, b, coef, poles, rinf=rinf)
+    if np.isscalar(z): return y[0]
+    y.shape = np.shape(z)
+    return y
+    
+def evalr_partialfraction(opinv, v, coef, poles, rinf=None, v0=None):
     if v0 is None:
         v0 = v+0j
-    if c0 is None:
+    if rinf is None:
         n = len(poles)
-        c0 = (-1)**n
-    y2 = c0*v0
+        rinf = (-1)**n
+    y2 = rinf*v0
     for (c,s) in zip(coef,poles):
         psi=opinv(s,v)
         y2 += c*psi
